JDTModule: fixed bug noted in #405 where jar module was not usable
This code is frankly a mess. @davidfestal should review my fix. From what I gather
the only place where a JDTModule is constructed it is never passed a null fragment roots
so I check for emptiness. Also comparing a String to a File didn't work.
I don't understand why there's a special case for jar files, or why we even go on checking
the rest of the fragments after we found the jar file, but hey…