Fix #1112 - format selection cuts off last character
In ceylon/ceylon.formatter@bbb920b, I made all formatter options
"writeable" by changing their string (Java toString()) to something that
can be used in a configuration file (profile); to support this for the
line break objects, I added a new attribute "text" holding the actual
line break text.

When the IDE used toString().length() to determine how much text to cut
off (the last line break has to be cut off), it actually measured the
length of "lf", not "\n". (On a Windows system, "crlf" would probably have
resulted in a cutoff of three characters, but I haven't tested that.)

The solution is of course trivial: We now need to use getText() instead
of toString().