Formatter config: Parse preview only once
We remember the lexed and parsed preview code, avoiding the need to
re-lex and -parse every time it needs to be updated. The preview now
updates almost instantly.

Removes CeylonContentFormatter because with the abstraction of IDocument
removed (I thought about a CeylonParsedDocument that would cache its
parse, but of course it should update that each time its text changes,
so we wouldn't gain anything), there's no point in keeping it.