Completely fix formatter config inlineAnnotations
The previous fix (677b021) wasn't enough - it still broke when the
option was 'all'. To truly fix this, we need to be able to disable the
text preference when it's created, before its listeners are attached.
(Some other tabs can apparently enable and disable preferences during
initialization without update problems - my best guess is that it's only
problematic for string properties, but I didn't investigate it too
much.)

Also, the handling of this "union" option was pretty broken. Now we
simply display 'all' in the text field when 'all' is checked (which makes
sense because that's what we'll write to the config file). When that
checkbox is unchecked, we insert the default inline annotations into the
textfield.

Also fixes the validity test. It's allowed to declare any annotations as
inline, so we shouldn't have a list of "allowed" annotations; instead,
we now only check if it's a valid annotation name.

Includes a workaround for ceylon/ceylon-spec#1106.