Fix formatter config spaceAfterParamListClosingParen
Different from 9d041de and 30ce055 because here, the int and the boolean
mean the same thing - the int is just a more precise way of expressing
it. So now we'll update the checkbox every time the int value changes,
and set the int value to the according end of the range (-80 / 80 -
somewhat arbitrary, but the true min / max values from ceylon.formatter
are way too big for normal use) every time the checkbox changes. So a 0
is saved as a 0, a -50 is saved as a -50, and an 80 is saved as 'true',
no matter if you arrived at it via number field or checkbox.