Fix #1443
findNode() now takes an additional token list (see
ceylon/ceylon-ide-common@5d509a0 for details), which we have to pass in
lots of places. Most of the time, we can get it from the
CeylonParseController; occasionally, we just pass null. (That's not
necessarily bad: for example, we don't have to worry about whitespace in
the selection when the selection comes from an error marker.)