Globally make completion non-blocking.
For #1589 (Run the "Preparing completions" task in a background thread)

Fixes also:
- #1814 : Very slow auto-completion work
- #1815 : Double dots

There were 2 main problems:

- the UI was frozen because we had to construct completions in another
thread, and run the `Display` event loop manually while waiting for the
result.
Doing this now allows registering a document-change listener that
cancels the completion search as soon as some new text is typed in the
editor.

- the images should not be retrieved during the completion construction
(this can causes a dead-lock), but lazily when
`CompletionProposal.getImage()` is called.

Additionally, this commit also enables retrieving partial fast
completion results when in auto-activation mode, or when a build is
running. Each time a partial completion result is returned, a warning is
shown in the status bar.