Fix #1844 (error with IntStream)
THis was due to the fact that the JDT model loader names the parameters
of Java methods according to their type. So the parameter was named
`int`, which led to an error in the backend since this is a Java
reserved keyword. I changed to prefix the parameter names with `_`.